requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
alembic==1.8.1
    # via flask-migrate
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via flask
commonmark==0.9.1
    # via rich
cryptography==38.0.3
    # via
    #   msal
    #   pyjwt
flask==2.2.2
    # via
    #   -r requirements.in
    #   flask-migrate
    #   flask-sqlalchemy
    #   flask-wtf
flask-migrate==3.1.0
    # via -r requirements.in
flask-sqlalchemy==3.0.2
    # via
    #   -r requirements.in
    #   flask-migrate
flask-wtf==1.0.1
    # via -r requirements.in
greenlet==2.0.1
    # via sqlalchemy
idna==3.4
    # via requests
importlib-metadata==5.0.0
    # via
    #   alembic
    #   flask
importlib-resources==5.10.0
    # via alembic
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via flask
mako==1.2.3
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
msal==1.20.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pygments==2.13.0
    # via rich
pyjwt[crypto]==2.6.0
    # via msal
pyodbc==4.0.34
    # via -r requirements.in
python-dotenv==0.21.0
    # via -r requirements.in
requests==2.28.1
    # via msal
rich==12.6.0
    # via -r requirements.in
sqlalchemy==1.4.43
    # via
    #   alembic
    #   flask-sqlalchemy
typing-extensions==4.4.0
    # via rich
urllib3==1.26.12
    # via requests
werkzeug==2.2.2
    # via flask
wtforms==3.0.1
    # via flask-wtf
zipp==3.10.0
    # via
    #   importlib-metadata
    #   importlib-resources
